+++ src/Accounts/Accounts/Authentication/LoginHandoffTracker.cs
@@ -35,22 +35,73 @@ namespace Sovereign.Accounts.Accounts.Authentication
     {
 
         /// <summary>
+        /// 
+        /// </summary>
+        private sealed class HandoffToken
+        {
+
+            /// <summary>
+            /// Account ID.
+            /// </summary>
+            private readonly Guid accountId;
+
+            /// <summary>
+            /// IP address from which the login request originated.
+            /// </summary>
+            private readonly string ipAddress;
+
+            public HandoffToken(Guid accountId, string ipAddress)
+            {
+                this.accountId = accountId;
+                this.ipAddress = ipAddress;
+            }
+
+            public override bool Equals(object o)
+            {
+                return o is HandoffToken token && this == token;
+            }
+
+            public static bool operator ==(HandoffToken left, HandoffToken right)
+            {
+                return (left == null && right == null) ||
+                        left != null && right != null &&
+                        left.accountId == right.accountId &&
+                        left.ipAddress.Equals(right.ipAddress);
+            }
+
+            public static bool operator !=(HandoffToken left, HandoffToken right)
+            {
+                return !(left == right);
+            }
+
+            public override int GetHashCode()
+            {
+                var hash = 17;
+                hash = hash * 31 + accountId.GetHashCode();
+                hash = hash * 31 + ipAddress.GetHashCode();
+                return hash;
+            }
+
+        }
+
+        /// <summary>
         /// Account IDs with pending handoffs.
         /// </summary>
-        private readonly ISet<Guid> pendingHandoffAccountIds
-            = new HashSet<Guid>();
+        private readonly ISet<HandoffToken> pendingHandoffs
+            = new HashSet<HandoffToken>();
 
         /// <summary>
         /// Adds an account to the pending handoff list.
         /// </summary>
         /// <param name="accountId">Account ID.</param>
+        /// <param name="ipAddress">IP address from which the login request originated.</param>
         /// <remarks>
         /// If the account is already in the pending handoff list, this
         /// method has no effect.
         /// </remarks>
-        public void AddPendingHandoff(Guid accountId)
+        public void AddPendingHandoff(Guid accountId, string ipAddress)
         {
-            pendingHandoffAccountIds.Add(accountId);
+            pendingHandoffs.Add(new HandoffToken(accountId, ipAddress));
         }
 
         /// <summary>
@@ -58,10 +109,11 @@ namespace Sovereign.Accounts.Accounts.Authentication
         /// present.
         /// </summary>
         /// <param name="accountId">Account ID.</param>
+        /// <param name="ipAddress">IP address from which the connection originated.</param>
         /// <returns>true if the account was in the list; false otherwise.</returns>
-        public bool TakePendingHandoff(Guid accountId)
+        public bool TakePendingHandoff(Guid accountId, string ipAddress)
         {
-            return pendingHandoffAccountIds.Remove(accountId);
+            return pendingHandoffs.Remove(new HandoffToken(accountId, ipAddress));
         }
 
     }